script/requirements.txt
Adafruit-DHT==1.4.0
mysql-connector==2.2.9
RPi.GPIO==0.6.5